+++ css/style.css
@@ -647,7 +647,7 @@ dialog {
 }
 
 .main-banner::before {
-	background-image: url("./images/blockchain_background.png");
+	background-image: url("./res/blockchain_background.png");
 	content: "";
 	background-size: cover;
 	position: absolute;
@@ -861,7 +861,7 @@ dialog {
 }
 
 .video-section {
-	background: url(images/video-bg.jpg) no-repeat fixed center;
+	background: url(res/video-bg.jpg) no-repeat fixed center;
 	background-size: cover;
 	overflow: hidden;
 }
@@ -992,12 +992,12 @@ dialog {
 	top: 0;
 	width: 100%;
 	height: 100%;
-	background: url(images/pattern.png);
+	background: url(res/pattern.png);
 }
 
 .bx-prev {
 	left: -90px !important;
-	background: url(images/arrow-left.png) no-repeat !important;
+	background: url(res/arrow-left.png) no-repeat !important;
 	background-size: auto auto !important;
 	background-size: 100% 100% !important;
 	color: #000000;
@@ -1006,7 +1006,7 @@ dialog {
 
 .bx-next {
 	right: -90px !important;
-	background: url(images/arrow-right.png) no-repeat !important;
+	background: url(res/arrow-right.png) no-repeat !important;
 	background-size: auto auto !important;
 	background-size: 100% 100% !important;
 	color: #000000;
@@ -1155,7 +1155,7 @@ input {
 	top: 0;
 	width: 100%;
 	height: 100%;
-	background: url(images/pattern.png);
+	background: url(res/pattern.png);
 }
 
 .download ul {
