+++ views/app.ejs
@@ -4,12 +4,12 @@
 <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     <title>INFT App</title>
-    <link rel="shortcut icon" href="./images/favicon.png" type="image/x-icon">
+    <link rel="shortcut icon" href="./res/favicon.png" type="image/x-icon">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/bxslider/4.2.12/jquery.bxslider.css">
-    <link rel="stylesheet" href="../css/animate.css">
+    <link rel="stylesheet" href="css/animate.css">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css"
         integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
-    <link rel="stylesheet" type="text/css" href="../style.css" />
+    <link rel="stylesheet" type="text/css" href="css/style.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css">
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -18,7 +18,7 @@
 <body>
     <!-- Hidden Spinning Wheel -->
     <div class="spin-background" id="spin">
-        <img class="spin" src="../images/icon.png">
+        <img class="spin" src="../res/icon.png">
     </div>
     <!-- End Spinning Wheel -->
 
@@ -41,7 +41,7 @@
             </button>
             <div id="sidenav" class="sidenav">
                 <a href="javascript:void(0)" class="closebtn" id="nav-closer">&times;</a>
-                <a href="/"><img class="wow fadeInUp app-logo" src="../images/logo.png" data-wow-delay="0.2s"></a>
+                <a href="/"><img class="wow fadeInUp app-logo" src="../res/logo.png" data-wow-delay="0.2s"></a>
                 <a href="#"><i></i></a>
                 <a href="#"><h4>Hi, <%- username %>!</h4></a>
                 <a href="#"><i></i></a>
@@ -79,7 +79,7 @@
                         var output = <%- userdata %>;
                     </script>
                     <div class="tile">
-                        <img src="../images/sample/1.jpg" class="tile-icon centered">
+                        <img src="../res/sample/1.jpg" class="tile-icon centered">
                         <script>
                             document.write("<h4>" + output.name + "</h4>");
                             document.write("<p>" + output.username + "</p>");
@@ -87,7 +87,7 @@
                         <a href="#" class="about-btn-solid read-more-nft">Read More</a>
                     </div>
                     <div class="tile">
-                        <img src="../images/sample/2.png" class="tile-icon">
+                        <img src="../res/sample/2.png" class="tile-icon">
                         <script>
                             document.write("<h4>" + output.name + "</h4>");
                             document.write("<p>" + output.username + "</p>");
@@ -95,7 +95,7 @@
                         <a href="#" class="about-btn-solid read-more-nft">Read More</a>
                     </div>
                     <div class="tile">
-                        <img src="../images/sample/3.png" class="tile-icon">
+                        <img src="../res/sample/3.png" class="tile-icon">
                         <script>
                             document.write("<h4>" + output.name + "</h4>");
                             document.write("<p>" + output.username + "</p>");
