+++ views/index.ejs
@@ -4,12 +4,12 @@
 <head>
   <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
   <title>INFT</title>
-  <link rel="shortcut icon" href="./images/favicon.png" type="image/x-icon">
+  <link rel="shortcut icon" href="./res/favicon.png" type="image/x-icon">
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/bxslider/4.2.12/jquery.bxslider.css">
   <link rel="stylesheet" href="css/animate.css">
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css"
     integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
-  <link rel="stylesheet" type="text/css" href="style.css" />
+  <link rel="stylesheet" type="text/css" href="css/style.css" />
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <!-- Font Google -->
@@ -20,7 +20,7 @@
 <body>
   <!-- Hidden Spinning Wheel -->
   <div class="spin-background" id="spin">
-    <img class="spin" src="../images/icon.png">
+    <img class="spin" src="../res/icon.png">
     <!-- <a href="/" class="x"><i class="fa-solid fa-x"></i></a> -->
   </div>
   <!-- End Spinning Wheel -->
@@ -38,7 +38,7 @@
   <!-- Navbar -->
   <nav class="navbar navbar-expand-lg">
     <div class="container">
-      <a class="navbar-brand navbar-logo mobile-hidden" href="/"> <img src="images/logo.png" alt="logo" class="logo-1">
+      <a class="navbar-brand navbar-logo mobile-hidden" href="/"> <img src="res/logo.png" alt="logo" class="logo-1">
       </a>
       <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
         aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"> <span
@@ -74,7 +74,7 @@
     <div>
       <div class="container centered-section">
         <div class="text-center">
-          <img src="images/logo-highres.png" class="wow fadeInUp header-padding-image" data-wow-delay="0.4s" />
+          <img src="res/logo-highres.png" class="wow fadeInUp header-padding-image" data-wow-delay="0.4s" />
           <h4>A powerful, customizable, and scalable CRM solution based on modern blockchain technologies.</h4>
           <div class="row">
             <div id="browse-docs" class="col-md-12"><a class="about-btn-solid">Browse Docs</a></div>
@@ -191,7 +191,7 @@
                 </div>
               </div>
               <div class="col-md-4 d-none d-md-block d-lg-block">
-                <div class="feature-mobile"> <img src="images/iphone-screen-with-shadow.png"
+                <div class="feature-mobile"> <img src="res/iphone-screen-with-shadow.png"
                     class="img-fluid wow fadeInUp" /> </div>
               </div>
               <div class="col-md-4 col-sm-12">
@@ -245,21 +245,21 @@
           <div class="section-content text-center">
             <div class="row">
               <div class="col-md-4">
-                <div class="team-detail wow bounce" data-wow-delay="0.2s"> <img src="images/user1.jpg"
+                <div class="team-detail wow bounce" data-wow-delay="0.2s"> <img src="res/user1.jpg"
                     class="img-fluid" />
                   <h4>Nitu Singh</h4>
                   <p>Marketing Specialist</p>
                 </div>
               </div>
               <div class="col-md-4">
-                <div class="team-detail wow bounce" data-wow-delay="0.4s"> <img src="images/user2.jpg"
+                <div class="team-detail wow bounce" data-wow-delay="0.4s"> <img src="res/user2.jpg"
                     class="img-fluid" />
                   <h4>Yogesh Singh</h4>
                   <p>CEO & Founder</p>
                 </div>
               </div>
               <div class="col-md-4">
-                <div class="team-detail wow bounce" data-wow-delay="0.6s"> <img src="images/user3.jpg"
+                <div class="team-detail wow bounce" data-wow-delay="0.6s"> <img src="res/user3.jpg"
                     class="img-fluid" />
                   <h4>Nisha Sharma</h4>
                   <p>Web Developer</p>
@@ -291,19 +291,19 @@
               <div class="offset-md-2 col-md-8 col-sm-12">
                 <div class="slider">
                   <div class="slider-item">
-                    <div class="test-img"><img src="images/user1.jpg" alt="Placeholder" width="157" height="157"></div>
+                    <div class="test-img"><img src="res/user1.jpg" alt="Placeholder" width="157" height="157"></div>
                     <div class="test-text"><span class="title"><span>John Michal</span> Digital Designer</span>
                       Consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna
                       aliquam eratvo lutpat.</div>
                   </div>
                   <div class="slider-item">
-                    <div class="test-img"><img src="images/user3.jpg" alt="Placeholder" width="157" height="157"></div>
+                    <div class="test-img"><img src="res/user3.jpg" alt="Placeholder" width="157" height="157"></div>
                     <div class="test-text"><span class="title"><span>Steve Smith</span> App User</span> Euismod
                       tincidunt ut laoreet dolore magna aliquam eratvo lutpat. Ut wisi enim ad minim veniam, quis
                       nostrud v</div>
                   </div>
                   <div class="slider-item">
-                    <div class="test-img"><img src="images/user3.jpg" alt="Placeholder" width="157" height="157"></div>
+                    <div class="test-img"><img src="res/user3.jpg" alt="Placeholder" width="157" height="157"></div>
                     <div class="test-text"><span class="title"><span>Gordon Shaw</span> Blogger</span> Consectetuer
                       adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam eratvo
                       lutpat. Ut wisi enim ad minim veniam, quis nostrud v</div>
@@ -453,8 +453,8 @@
         <div class="col-md-12">
           <div class="section-content text-center">
             <ul>
-              <li><a href="#"><img src="images/appstore.png" class="wow fadeInUp" data-wow-delay="0.4s" /></a></li>
-              <li><a href="#"><img src="images/playstore.png" class="wow fadeInUp" data-wow-delay="0.7s" /></a></li>
+              <li><a href="#"><img src="res/appstore.png" class="wow fadeInUp" data-wow-delay="0.4s" /></a></li>
+              <li><a href="#"><img src="res/playstore.png" class="wow fadeInUp" data-wow-delay="0.7s" /></a></li>
             </ul>
           </div>
         </div>
@@ -531,12 +531,12 @@
       navbar_logo = $(".navbar-logo");
 
       if (bodyScroll > 50) {
-        $('.navbar-logo img').attr('src', 'images/logo-black.png');
+        $('.navbar-logo img').attr('src', 'res/logo-black.png');
         navbar.addClass("nav-scroll");
         navbar_logo.removeClass("mobile-hidden");
 
       } else {
-        $('.navbar-logo img').attr('src', 'images/logo.png');
+        $('.navbar-logo img').attr('src', 'res/logo.png');
         navbar.removeClass("nav-scroll");
         navbar_logo.addClass("mobile-hidden");
       }
@@ -547,10 +547,10 @@
         navbar = $(".navbar");
 
       if (bodyScroll > 50) {
-        $('.navbar-logo img').attr('src', 'images/logo-black.png');
+        $('.navbar-logo img').attr('src', 'res/logo-black.png');
         navbar.addClass("nav-scroll");
       } else {
-        $('.navbar-logo img').attr('src', 'images/logo-white.png');
+        $('.navbar-logo img').attr('src', 'res/logo-white.png');
         navbar.removeClass("nav-scroll");
       }
