+++ views/login.ejs
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="../css/animate.css">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css"
         integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
-    <link rel="stylesheet" type="text/css" href="../style.css" />
+    <link rel="stylesheet" type="text/css" href="css/style.css" />
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.1.0/css/all.css"
         integrity="sha384-lKuwvrZot6UHsBSfcMvOkWwlCMgc0TaWr+30HWe3a4ltaBwTZhyTEggF5tJv8tbt" crossorigin="anonymous">
 
@@ -21,7 +21,7 @@
 <body>
     <!-- Hidden Spinning Wheel -->
     <div class="spin-background" id="spin">
-        <img class="spin" src="../images/icon.png">
+        <img class="spin" src="../res/icon.png">
     </div>
     <!-- End Spinning Wheel -->
 
@@ -37,7 +37,7 @@
 
     <section class="main-banner-gradient">
         <div class="text-center">
-            <a href="/"><img class="header-padding-image wow fadeInUp" src="../images/logo-highres.png"
+            <a href="/"><img class="header-padding-image wow fadeInUp" src="../res/logo-highres.png"
                     data-wow-delay="0.2s"></a>
             <form class="centered prelative normal-padding" id="signin_form" method="post">
                 <input type="text" id="username" class="form-input w-75" name="username"
@@ -53,11 +53,11 @@
             <span class="line"></span>
             <div class="" id="signinmethods">
                 <div class="container normal-padding prelative">
-                    <a id="ms" href="/signin/microsoft"><img src="../images/ms_login.png" class="wow fadeInUp prelative"
+                    <a id="ms" href="/signin/microsoft"><img src="../res/ms_login.png" class="wow fadeInUp prelative"
                             data-wow-delay="0.6s" /></a>
                 </div>
                 <div class="container normal-padding prelative">
-                    <a id="mm" href="#"><img src="../images/metamask_login.png" class="wow fadeInUp prelative"
+                    <a id="mm" href="#"><img src="../res/metamask_login.png" class="wow fadeInUp prelative"
                             data-wow-delay="0.8s" /></a>
                 </div>
             </div>
